requirements.txt
openenv-core[core]>=0.2.0
pydantic>=2.5.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.25.2
